pyproject.toml
[tool.poetry]
name = "weya-takehome"
version = "0.1.0"
description = ""
authors = ["Gupta, Depika"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.9.0"
chainlit = "^1.1.301"
speechrecognition = "^3.10.4"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
pydub = "^0.25.1"
elevenlabs = "^1.3.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
